+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/tracker/export/event/EventRequestParams.java
@@ -142,6 +142,8 @@ public class EventRequestParams implements PageRequestParams {
 
   private EventStatus status;
 
+  private String attachment;
+
   /**
    * @deprecated use {@link #attributeCategoryCombo}
    */
